Initial setup for the generic model decoder. Cleaning up assimp implementation.
